fix: validacao de evento existente para deleção de Local e outras correções

+++ src/main/java/unicv/poo/eventos_api/service/LocalService.java
@@ -1,5 +1,6 @@
 package unicv.poo.eventos_api.service;
 
+import lombok.RequiredArgsConstructor;
 import unicv.poo.eventos_api.entity.Local;
 import unicv.poo.eventos_api.exception.RegraNegocioException;
 import unicv.poo.eventos_api.dto.LocalRequestDTO;
@@ -7,7 +8,6 @@ import unicv.poo.eventos_api.dto.LocalResponseDTO;
 import unicv.poo.eventos_api.mapper.LocalMapper;
 import unicv.poo.eventos_api.repository.EventoRepository;
 import unicv.poo.eventos_api.repository.LocalRepository;
-import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.lang.NonNull;
 import org.springframework.stereotype.Service;
 
@@ -16,15 +16,13 @@ import jakarta.persistence.EntityNotFoundException;
 import java.util.List;
 
 @Service
+@RequiredArgsConstructor
 public class LocalService {
 
-    @Autowired
-    private LocalRepository localRepository;
-
-    @Autowired
-     private EventoRepository eventoRepository; 
-    @Autowired
-    private LocalMapper localMapper;
+    private final LocalRepository localRepository;
+    private final EventoRepository eventoRepository;
+    private final EventoService eventoService;
+    private final LocalMapper localMapper;
 
     public List<LocalResponseDTO> listarTodos(){
         List<Local> locais = localRepository.findAll();
@@ -45,12 +43,12 @@ public class LocalService {
 
     public LocalResponseDTO atualizar(Long id, LocalRequestDTO localRequestDTO){
         Local localExistente = localRepository.findById(id)
-        .orElseThrow(() -> new EntityNotFoundException("Não é possível atualizar: Local não encontrado."));
+        .orElseThrow(() -> new EntityNotFoundException("Local não encontrado."));
 
         if(localRequestDTO.capacidade()< localExistente.getCapacidade()){
         boolean possuiEventoMaior = eventoRepository.existsByLocalIdAndCapacidadeGreaterThan(id, localRequestDTO.capacidade());
             if(possuiEventoMaior){
-                throw new RegraNegocioException("Não é possível reduzir a capacidade: existe um evento previsto que excede a nova capacidade.");
+                throw new RegraNegocioException("Existe um evento previsto que excede a nova capacidade.");
             }
         }
 
@@ -64,8 +62,13 @@ public class LocalService {
 
     public void deletar(@NonNull Long id){
         if(!localRepository.existsById(id)){
-            throw new EntityNotFoundException("Não é possivel remover: Local não encontrado.");
+            throw new EntityNotFoundException("Local não encontrado.");
         }
+
+        if (eventoService.existeEventoNoLocal(id)) {
+            throw new RegraNegocioException("Existem eventos cadastrados para este local.");
+        }
+
         localRepository.deleteById(id);
     }
 }
